requirements.txt
flask>=3.0,<4.0
pydantic>=2.7,<3.0

# Compatible set (langchain 1.2.x expects langgraph >=1.0.2)
langchain>=1.2.0,<1.3.0
langgraph>=1.0.2,<1.1.0
langchain-openai>=1.1.0,<1.2.0

python-dotenv>=1.0,<2.0
